requirements.txt
streamlit
streamlit-webrtc
pydub
openai-whisper
torch
numpy
av